fix the setup-and-run.bat and we are using pm2
