+++ README.md
@@ -1,0 +1,1 @@
+please install node-installer.msi first
