--- gui-proto/Dockerfile
@@ -1,19 +1,0 @@
-﻿# Use the official Python image from the Docker Hub
-FROM python:3.12-slim
-
-USER root
-
-# Set the working directory in the container
-WORKDIR /app
-
-# Copy the requirements file into the container
-COPY requirements.txt .
-
-# Install the dependencies
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy the rest of the application code into the container
-COPY . .
-
-# Set the default command to run the application
-CMD ["streamlit", "run", "main.py", "--server.port", "8081"]
