--- gui-proto/docker-compose.yml
@@ -1,8 +1,0 @@
-services:
-  proteus-ui:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-      - "8081:8081"
-    restart: always
