--- gui-proto/notes.md
@@ -1,3 +1,0 @@
-﻿- error on config.toml
-  - locate c/users/username/.streamlit/<something>.toml and delete it
-  - remove BOM from all toml files (endianness byte added by Windows os)
