--- gui-proto/requirements.txt
@@ -1,7 +1,0 @@
-streamlit==1.44.0
-streamlit-elements
-pyyaml
-pandas
-numpy
-altair
-xlsxwriter
