--- gui-proto/tests/data/test_config.yml
@@ -1,6 +1,0 @@
-﻿theme:
-  base: light
-  primaryColor: "#F63366"
-  backgroundColor: "#F0F2F6"
-  secondaryBackgroundColor: "#FFFFFF"
-  textColor: "#262730"
