+++ proteus-ui/notes.md
@@ -1,0 +1,3 @@
+﻿- error on config.toml
+  - locate c/users/username/.streamlit/<something>.toml and delete it
+  - remove BOM from all toml files (endianness byte added by Windows os)
