+++ software_update/__main__.py
@@ -1,0 +1,1 @@
+print("im from software update")
